+++ apps/web/src/hooks/auction/usePostAuction.tsx
@@ -1,0 +1,3 @@
+const usePostAuction = () => {};
+
+export default usePostAuction;
